Добавлены противники

+++ Shoot'em up/Assets/Scripts/Player/Config/PlayerConfig.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class PlayerConfig : ScriptableObject
 {
     [field: SerializeField] public float MoveSpeed { get; private set; }
-    [field: SerializeField] public int MaxHP { get; private set; }
-    [field: SerializeField] public int MinHP { get; private set; }
+    [field: SerializeField] public float MaxHP { get; private set; }
+    [field: SerializeField] public float MinHP { get; private set; }
 
 }
